refactor[Comp]: Move AppBar to separate component
