+++ src/components/AppBar/appbar.module.scss
@@ -1,0 +1,15 @@
+.logo {
+  margin-right: 8px;
+}
+
+.logo-text {
+  flex-grow: 1;
+}
+
+.menu-appbar {
+  margin-top: 50px;
+}
+
+.logout-icon {
+  color: var(--mui-palette-error-main);
+}
