+++ src/components/BaseLayout/baselayout.module.scss
@@ -2,22 +2,6 @@
   display: flex;
 }
 
-.logo {
-  margin-right: 8px;
-}
-
-.logo-text {
-  flex-grow: 1;
-}
-
-.menu-appbar {
-  margin-top: 50px;
-}
-
-.logout-icon {
-  color: var(--mui-palette-error-main);
-}
-
 .main-container {
   max-width: 0px;
 }
